# openhwy/platforms/fed-tms/apps/api/openapi.yaml
openapi: 3.0.3
info:
  title: FED-TMS API
  description: Fast & Easy Dispatching - Transportation Management System API
  version: 1.0.0
  contact:
    name: FED-TMS Support
  license:
    name: Proprietary
servers:
  - url: http://localhost:3002/api/v1
    description: Development server
  - url: https://api.fedtms.com/api/v1
    description: Production server
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Error:
      type: object
      required:
        - success
        - error
        - timestamp
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
          example: Resource not found
        code:
          type: string
          example: NOT_FOUND
        timestamp:
          type: string
          format: date-time
    Pagination:
      type: object
      required:
        - total
        - page
        - limit
        - totalPages
      properties:
        total:
          type: integer
          example: 100
        page:
          type: integer
          example: 1
        limit:
          type: integer
          example: 10
        totalPages:
          type: integer
          example: 10

    # ==================== AUTH SCHEMAS ====================
...
    User:
      type: object
      required:
        - id
        - email
        - first_name
        - last_name
      properties:
        id:
          type: string
        email:
          type: string
          format: email
        first_name:
          type: string
        last_name:
          type: string
        phone:
          type: string
        avatar_url:
          type: string
          format: uri
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    UserWithRoles:
      allOf:
        - $ref: '#/components/schemas/User'
        - type: object
          properties:
            roles:
              type: array
              items:
                type: object
                properties:
                  id:
                    type: string
                  name:
                    type: string
...
    Conversation:
      type: object
      required:
        - id
        - company_id
      properties:
        id:
          type: string
        company_id:
          type: string
        name:
          type: string
        is_group:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    ConversationDetail:
      allOf:
        - $ref: '#/components/schemas/Conversation'
        - type: object
          properties:
            participants:
              type: array
              items:
                type: object
                properties:
                  user_id:
                    type: string
                  user:
                    type: object
                    properties:
                      id:
                        type: string
                      email:
                        type: string
                      first_name:
                        type: string
                      last_name:
                        type: string
            messages:
              type: array
              items:
                $ref: '#/components/schemas/Message'
    Message:
      type: object
      required:
        - id
        - conversation_id
        - sender_id
        - content
      properties:
        id:
          type: string
        conversation_id:
          type: string
        sender_id:
          type: string
        content:
          type: string
        message_type:
          type: string
          enum: [text, image, file, system]
        attachment_url:
          type: string
          format: uri
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    CreateConversationRequest:
      type: object
      properties:
        name:
          type: string
        is_group:
          type: boolean
          default: false
        participant_ids:
          type: array
          items:
            type: string
    SendMessageRequest:
      type: object
      required:
        - content
      properties:
        content:
          type: string
          minLength: 1
        message_type:
          type: string
          enum: [text, image, file, system]
          default: text
        attachment_url:
          type: string
          format: uri

    # ==================== RESPONSE WRAPPERS ====================
...
paths:
...
  /users:
    get:
      summary: List users
      description: Get all users in company
      tags:
        - Users
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
        - name: search
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Users list
          content:
            application/json:
              schema:
                allOf:
                  - type: object
                    properties:
                      success:
                        type: boolean
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/User'
                      pagination:
                        $ref: '#/components/schemas/Pagination'
                      timestamp:
                        type: string
                        format: date-time

    post:
      summary: Create user
      description: Create new user in company
      tags:
        - Users
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
                - first_name
                - last_name
                - role_id
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
                  minLength: 8
                first_name:
                  type: string
                last_name:
                  type: string
                phone:
                  type: string
                role_id:
                  type: string
      responses:
        '201':
          description: User created
          content:
            application/json:
              schema:
                allOf:
                  - type: object
                    properties:
                      success:
                        type: boolean
                      data:
                        $ref: '#/components/schemas/User'
                      timestamp:
                        type: string
                        format: date-time
  /users/{id}:
    get:
      summary: Get user by ID
      tags:
        - Users
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: User details
          content:
            application/json:
              schema:
                allOf:
                  - type: object
                    properties:
                      success:
                        type: boolean
                      data:
                        $ref: '#/components/schemas/UserWithRoles'
                      timestamp:
                        type: string
                        format: date-time
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    put:
      summary: Update user
      tags:
        - Users
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                first_name:
                  type: string
                last_name:
                  type: string
                phone:
                  type: string
      responses:
        '200':
          description: User updated
          content:
            application/json:
              schema:
                allOf:
                  - type: object
                    properties:
                      success:
                        type: boolean
                      data:
                        $ref: '#/components/schemas/User'
                      timestamp:
                        type: string
                        format: date-time

    delete:
      summary: Delete user
      tags:
        - Users
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: User deleted
          content:
            application/json:
              schema:
                allOf:
                  - type: object
                    properties:
                      success:
                        type: boolean
                      data:
                        type: object
                        properties:
                          message:
                            type: string

  # ==================== DRIVERS ====================
...
  /conversations:
    get:
      summary: List conversations
      tags:
        - Conversations
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
      responses:
        '200':
          description: Conversations list
          content:
            application/json:
              schema:
                allOf:
                  - type: object
                    properties:
                      success:
                        type: boolean
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/Conversation'
                      pagination:
                        $ref: '#/components/schemas/Pagination'
                      timestamp:
                        type: string
                        format: date-time

    post:
      summary: Create conversation
      tags:
        - Conversations
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateConversationRequest'
      responses:
        '201':
          description: Conversation created
          content:
            application/json:
              schema:
                allOf:
                  - type: object
                    properties:
                      success:
                        type: boolean
                      data:
                        $ref: '#/components/schemas/Conversation'
                      timestamp:
                        type: string
                        format: date-time
  /conversations/{id}:
    get:
      summary: Get conversation by ID
      tags:
        - Conversations
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Conversation details
          content:
            application/json:
              schema:
                allOf:
                  - type: object
                    properties:
                      success:
                        type: boolean
                      data:
                        $ref: '#/components/schemas/ConversationDetail'
                      timestamp:
                        type: string
                        format: date-time

    put:
      summary: Update conversation
      tags:
        - Conversations
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
      responses:
        '200':
          description: Conversation updated
          content:
            application/json:
              schema:
                allOf:
                  - type: object
                    properties:
                      success:
                        type: boolean
                      data:
                        $ref: '#/components/schemas/ConversationDetail'
                      timestamp:
                        type: string
                        format: date-time

    delete:
      summary: Delete conversation
      tags:
        - Conversations
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Conversation deleted
          content:
            application/json:
              schema:
                allOf:
                  - type: object
                    properties:
                      success:
                        type: boolean
                      data:
                        type: object
                        properties:
                          message:
                            type: string
  /conversations/{id}/messages:
    get:
      summary: List messages in conversation
      tags:
        - Conversations
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
      responses:
        '200':
          description: Messages list
          content:
            application/json:
              schema:
                allOf:
                  - type: object
                    properties:
                      success:
                        type: boolean
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/Message'
                      pagination:
                        $ref: '#/components/schemas/Pagination'
                      timestamp:
                        type: string
                        format: date-time

    post:
      summary: Send message
      tags:
        - Conversations
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SendMessageRequest'
      responses:
        '201':
          description: Message sent
          content:
            application/json:
              schema:
                allOf:
                  - type: object
                    properties:
                      success:
                        type: boolean
                      data:
                        $ref: '#/components/schemas/Message'
                      timestamp:
                        type: string
                        format: date-time